Add pytest and routes

+++ magalu/employee/views.py
@@ -1,37 +1,54 @@
 from rest_framework.response import Response
+from rest_framework.decorators import api_view
 from rest_framework.views import APIView
 from rest_framework.generics import get_object_or_404
+from rest_framework import status
 
 from .models import Employee
 from .serializers import EmployeeSerializer
 
-class EmployeeView(APIView):
-    def get(self, request):
+@api_view(['GET', 'DELETE', 'PUT'])
+def get_delete_update_employee(request, pk):
+    try:
+        employee = Employee.objects.get(pk=pk)
+    except Employee.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    # get details of a single employee
+    if request.method == 'GET':
+        serializer = EmployeeSerializer(employee)
+        return Response(serializer.data)
+    # delete a single employee
+    elif request.method == 'DELETE':
+        employee.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    # update details of a single employee
+    elif request.method == 'PUT':
+        serializer = EmployeeSerializer(employee, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET', 'POST'])
+def get_post_employee(request):
+    # get all employees
+    if request.method == 'GET':
         employees = Employee.objects.all()
         serializer = EmployeeSerializer(employees, many=True)
-        return Response({"employees": serializer.data})
-
-    def post(self, request):
-        employee = request.data.get('employee')
-
-        # Create an employee from the above data
-        serializer = EmployeeSerializer(data=employee)
-        if serializer.is_valid(raise_exception=True):
-            employee_saved = serializer.save()
-        return Response({"success": "Employee '{}' created successfully".format(employee_saved.title)})
-
-    def put(self, request, pk):
-        saved_employee = get_object_or_404(Employee.objects.all(), pk=pk)
-        data = request.data.get('employee')
-        serializer = EmployeeSerializer(instance=saved_employee, data=data, partial=True)
-
-        if serializer.is_valid(raise_exception=True):
-            employee_saved = serializer.save()
-        return Response({"success": "Employee '{}' updated successfully".format(employee_saved.title)})
-
-
-    def delete(self, request, pk):
-        # Get object with this pk
-        employee = get_object_or_404(Employee.objects.all(), pk=pk)
-        employee.delete()
-        return Response({"message": "Employee with id `{}` has been deleted.".format(pk)},status=204)
+        return Response(serializer.data)
+    # insert a new record for a employee
+    elif request.method == 'POST':
+        print(request)
+        data = {
+            'first_name': request.data.get('first_name'),
+            'last_name': request.data.get('last_name'),
+            'email': request.data.get('email'),
+            'department': request.data.get('department')
+        }
+        serializer = EmployeeSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
